backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.1
greenlet>=3.0.0

# Validation & Settings
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Telegram
python-telegram-bot>=20.7
aiogram>=3.3.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utils
httpx>=0.26.0
python-dotenv>=1.0.0
